// source/frontend/src/models/api/generated/ApiGen_Concepts_FinancialCode.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDateTime } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_FinancialCodeTypes } from './ApiGen_Concepts_FinancialCodeTypes';

// LINK: @backend/apimodels/Models/Concepts/FinancialCode/FinancialCodeModel.cs
export interface ApiGen_Concepts_FinancialCode extends ApiGen_Base_BaseAudit {
  id: number;
  type: ApiGen_Concepts_FinancialCodeTypes;
  code: string | null;
  description: string | null;
  displayOrder: number | null;
  effectiveDate: UtcIsoDateTime;
  expiryDate: UtcIsoDateTime | null;
}